\name{extractoR.extract-package}
\alias{extractoR.extract-package}
\alias{extractoR.extract}
\docType{package}
\encoding{UTF-8}
\title{CRAN packages metadata extractor (extract)}
\description{
  extractoR.extract is part of extractoR which is a set of R packages
  that can be used to fetch, extract and query CRAN packages
  metadata. extractoR.extract extracts packages metadata in dataframes.
}

\author{
Maëlick Claes <himself@maelick.net> & Prateek Rastogi <p.rastogi@outlook.com>

Maintainer: Maëlick Claes <himself@maelick.net>
}
