\name{extractoR.extract-package}
\alias{extractoR.extract-package}
\alias{extractoR.extract}
\docType{package}
\encoding{UTF-8}
\title{R packages metadata extractor (extract)}
\description{
  extractoR.extract is part of extractoR which is a set of R packages
  that can be used to fetch, extract and query R packages metadata.
  extractoR.extract extracts packages metadata.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\examples{
\dontrun{
}
}