\name{extractoR.extract-package}
\alias{extractoR.extract-package}
\alias{extractoR.extract}
\docType{package}
\title{CRAN packages metadata extractor (extract)}
\description{
  extractoR.extract is part of extractoR which is a set of R packages
  that can be used to fetch, extract and dump CRAN packages metadata in
  a SQL database. extractoR.extract extracts packages metadata in
  dataframes.
}
\details{
\tabular{ll}{
Package: \tab extractoR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-21\cr
License: \tab GPL-3\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Maëlick Claes <himself@maelick.net> & Prateek Rastogi <p.rastogi@outlook.com>

Maintainer: Maëlick Claes <himself@maelick.net>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
